requirements.txt
# LightRAG 项目依赖
# Python >= 3.10
# 硬件: 8x NVIDIA RTX 4090 (24GB)

# ========== 核心 RAG 框架 ==========
lightrag>=0.1.0

# ========== LLM 服务 (vLLM) ==========
vllm>=0.6.0               # 高性能推理引擎 (8x4090 优化)
openai>=1.0.0             # OpenAI 兼容 API

# ========== 文档提取 ==========
PyMuPDF>=1.23.0           # PDF 提取 (最快)
python-docx>=1.1.0        # DOCX 读取
docx2txt>=0.8             # DOC 读取

# ========== Embedding 模型 ==========
sentence-transformers>=2.2.0  # bge-m3
transformers>=4.35.0          # HuggingFace transformers

# ========== 科学计算 ==========
numpy>=1.24.0

# ========== 数据存储 ==========
aiofiles>=23.0.0
httpx>=0.24.0

# ========== 可选依赖 ==========
rapidocr-onnxruntime>=1.3.0   # OCR 支持（扫描 PDF）
chardet>=5.0.0                # 编码自动检测（文本文件）
# tenacity>=8.2.0              # 重试机制
# tiktoken>=0.5.0              # Token 计算
# redis>=5.0.0                 # 缓存层
# fastapi>=0.100.0             # API 服务
# uvicorn>=0.23.0              # ASGI 服务器
